refactor: GroupSpecification 패키지 이동
